chore: rename UnitsView to UnitManagerView
